Removed import/export of getHeaders and moved mainWrapper to a different file
