+++ src/index.js
@@ -1,10 +1,6 @@
 // import { parseFoodComposition } from '../projects2.0/FoodComposition/parser';
-import path from 'path';
-
 import csvToJson from './csvToJson';
 
-// https://github.com/GroceriStar/food-datasets-csv-parser/issues/180
-import getHeaders from './getHeaders';
 import parseCsv from './parseCsv';
 
 // @TODO i don't like this name,
@@ -15,37 +11,10 @@ import parseDirectoryFiles from './fileSystem';
 
 // i adding utils here only because it's my guess
 import { readAllFiles, joinPath, getList } from './utils';
-
-const mainWrapper = async (filename) => {
-  const headers = await getHeaders(filename);
-
-  // console.log(headers);
-  // I might have to include lines 5-14 insinde of an async function and await it.
-  // This will become clearer once parseDrectoryFiles is functional.
-
-  // @TODO this line is a good idea. For all of our csv parsers
-  // projects we have a separated folder with files + parser inside.
-  // we can make a root directory as default inside of `parserDirectoryFiles`
-  // and change it if passed another variable
-  const directory = './';
-
-  const directoryPath = path.join(__dirname, directory + filename);
-
-  // test version
-  // const directoryPath = path.join(__dirname, filename);
-
-  // console.log(filename);
-
-  console.log(directoryPath);
-  console.log('---');
-  console.log(headers);
-  // parseDirectoryFiles(directoryPath, headers);
-  // parseDirectoryFiles(filename, headers);
-};
+import mainWrapper from './mainWrapper';
 
 export {
   csvToJson,
-  getHeaders,
   // @TODO let's debate should we export
   // parseCsv or we just using it inside of csvToJson
   parseCsv,
